Created schedule page and home page, added route for the schedule page. Fixed header alignment in chore table.
